Change variable name for --prefix to ${SYSROOT}
It was a bit confusing when I was first reviewing the build scripts to see a variable ${LIBS} used in the way it is.  ${LIBS} usually contains the libraries to link with.  The common name for a '--prefix' path is the system root, or SYSROOT.  Using this terminology might save some others who look at this code a few moments confusion.